+++ test/callback-api.test.js
@@ -78,6 +78,14 @@ describe('Callback API', function() {
     }, 250);
   });
 
+  it('should ignore disabled tasks', function(done) {
+    p = pipsqueak({ task: task, disabled: true, interval: '100ms', }).start();
+    setTimeout(function() {
+      assert.equal(executions, 0);
+      done();
+    }, 250);
+  });
+
   it('should emit begin and end events', function(done) {
     var events = [];
     var handler = function(event) {